requirements.txt
# FastAPI и зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

# Векторные базы данных и эмбеддинги
faiss-cpu>=1.9.0
sentence-transformers==3.1.1
transformers>=4.38.0
torch>=2.0.0
numpy>=1.24.0

# Работа с Excel
openpyxl==3.1.2
pandas>=2.0.0

# HTTP клиенты
httpx>=0.27.0,<1.0.0

# Мониторинг и логирование
structlog==24.4.0
python-json-logger==2.0.7

# Конфигурация и переменные окружения
python-dotenv==1.0.1
pydantic-settings==2.6.1

# Кэширование
aiocache==0.12.2

# Утилиты
tqdm==4.67.1

# Линтеры и форматтеры
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Дополнительные линтеры для FastAPI
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
pylint==3.0.3

# Для работы с типами
types-requests==2.31.0.10

# Дополнительные инструменты разработки
pre-commit==3.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.14.0
pytest-xdist==3.6.0
bandit==1.7.5
safety>=2.3.5

# Дополнительные инструменты
mypy-extensions==1.1.0
types-python-dateutil==2.9.0.20250822
